Back-end/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Authlib==0.11
Automat==20.2.0
beautifulsoup4==4.8.2
blinker==1.4
bs4==0.0.1
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
dnspython==1.16.0
filelock==3.0.12
Flask==1.0.2
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.5
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemy==0.56
gcloud==0.17.0
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jws==0.1.3
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.2
oauth2==1.9.0.post1
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==1.0.3
parsel==1.5.2
passlib==1.7.2
pathtools==0.1.2
Pillow==7.1.1
pipenv==2020.6.2
Protego==0.1.16
protobuf==3.6.1
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.6.6
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.0
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2019.3
queuelib==1.5.0
requests==2.20.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
scrapy-user-agents==0.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.15
sqlparse==0.3.1
Twisted==20.3.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.1
virtualenv==20.0.15
virtualenv-clone==0.5.4
w3lib==1.21.0
watchdog==0.10.2
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
Whoosh==2.7.4
zope.interface==5.0.2
